refactor: move some structs and
add functionality for saving data on external subtitle tracks
